refactor: extract community guidelines hook
